--- a/src/backends/tensorrt/plan_backend.cc
+++ b/src/backends/tensorrt/plan_backend.cc
@@ -337,9 +337,9 @@ PlanBackend::CreateExecutionContext(
     }
   }
 
-  // As we have visited all the input bindings at this point, if any of the
-  // shapes had dynamic dimension then is_dynamic_ flag would be set.
-  if (!context->is_dynamic_ &&
+  // Validate the batch dimension against the implicit batch dimension if
+  // available
+  if (context->engine_->hasImplicitBatchDimension() &&
       (context->max_batch_size_ > context->engine_->getMaxBatchSize())) {
     return Status(
         RequestStatusCode::INVALID_ARG,
@@ -985,12 +985,13 @@ PlanBackend::Context::Run(
     LOG_VERBOSE(1) << "Context with profile " << citr->second.profile_name_
                    << " [" << std::to_string(citr->first)
                    << "] is being executed for " << name_;
-    if (is_dynamic_) {
-      if (!citr->second.context_->allInputDimensionsSpecified()) {
-        return Status(
-            RequestStatusCode::INTERNAL,
-            "failed to specify the dimensions of all input bindings");
-      }
+    if (is_dynamic_ &&
+        (!citr->second.context_->allInputDimensionsSpecified())) {
+      return Status(
+          RequestStatusCode::INTERNAL,
+          "failed to specify the dimensions of all input bindings");
+    }
+    if (!engine_->hasImplicitBatchDimension()) {
       if (!citr->second.context_->enqueueV2(
               buffer_bindings_.data(), stream_, nullptr)) {
         cudaStreamSynchronize(stream_);