**NOTE** : Merge to master when TRT 7  is released.

When using fixed shape models with explicit batch dimensions there were a bunch of warnings being generated by TRT.

**Before:**
```
# ../clients/perf_client -m plan_nobatch_float32_float32_float32
*** Measurement Settings ***
  Batch size: 1
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
W1213 00:17:06.015177 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.016012 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.016669 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.017329 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.018013 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.018661 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.019374 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
W1213 00:17:06.019979 29518 logging.cc:46] Explicit batch network detected and batch size specified, use enqueue without batch size instead.
.
.
.

```

**Now**

```
 # ../clients/perf_client -m plan_nobatch_float32_float32_float32
 *** Measurement Settings ***
   Batch size: 1
   Measurement window: 5000 msec
   Using synchronous calls for inference
   Stabilizing using average latency
 
 Request concurrency: 1
   Client: 
     Request count: 11918
     Throughput: 2383.6 infer/sec
     Avg latency: 417 usec (standard deviation 99 usec)
     p50 latency: 423 usec
     p90 latency: 518 usec
     p95 latency: 566 usec
     p99 latency: 705 usec
     Avg HTTP time: 418 usec (send/recv 62 usec + response wait 356 usec)
   Server: 
     Request count: 14241
     Avg request latency: 107 usec (overhead 12 usec + queue 16 usec + compute 79 usec)
 
 Inferences/Second vs. Client Average Batch Latency
 Concurrency: 1, throughput: 2383.6 infer/sec, latency 417 usec
```